move mem structs into their own dir
