+++ src/process_module.rs
@@ -68,4 +68,9 @@ impl ProcessModule
             self.memory = buffer;
         }
     }
+
+    pub fn get_exports()
+    {
+
+    }
 }
